Using executor w/ queue so events are not dropped
